refactor: extract password change logic to user service and shared component
