+++ src/app/staff/settings/profile/page.tsx
@@ -11,8 +11,18 @@ import { StaffProfileEdit } from "@/components/staff/profile/staff-profile-edit"
  */
 export default function AdminSettingsProfilePage(): JSX.Element {
   return (
-    <Box sx={{ px: { xs: 2, md: 4 }, pt: { xs: 1, md: 1.5 }, pb: 4 }}>
-      <StaffProfileEdit />
+    <Box
+      sx={{
+        flex: 1,
+        overflowY: "auto",
+        minHeight: 0,
+        display: "flex",
+        flexDirection: "column",
+      }}
+    >
+      <Box sx={{ px: { xs: 2, md: 4 }, pt: { xs: 1, md: 1.5 }, pb: 4 }}>
+        <StaffProfileEdit />
+      </Box>
     </Box>
   );
 }
