+++ src/components/shared/index.ts
@@ -1,4 +1,5 @@
 export { AsyncContent } from "./async-content";
+export { ChangePasswordSection } from "./change-password-section";
 export { ConfirmDialog } from "./confirm-dialog";
 export { DetailField } from "./detail-field";
 export { ModalTitleBar } from "./modal-title-bar";
